btrc: consolidate recordings scope implementation
