+++ pkg/services/recordings/internal/lifecycle_capability.go
@@ -1,9 +1,16 @@
 package internal
 
 import (
+	"context"
+	"crypto/sha256"
+	"encoding/hex"
 	"errors"
-
+	"fmt"
 	"github.com/portpowered/infinite-you/pkg/services/recordings"
+	"github.com/portpowered/infinite-you/pkg/services/recordings/internal/canonical"
+	"strings"
+	"sync"
+	"time"
 )
 
 var _ recordings.RecordingLifecycle = (*combinedService)(nil)
@@ -217,4 +224,424 @@ func translateLifecycleError(err error) error {
 		kind = recordings.LifecycleErrorInvalidTerminalMetadata
 	}
 	return &recordings.LifecycleError{Kind: kind, Message: err.Error(), Cause: err}
+}
+
+type recordingScopeBinding struct {
+	recordingID recordings.RecordingID
+	eventScope  recordings.CanonicalEventScope
+	historical  bool
+	replayPlans map[recordings.ReplayPlanHandle]struct{}
+
+	mu          sync.Mutex
+	finalized   bool
+	closed      bool
+	terminal    recordings.RecordingScopeStatus
+	finalizeErr error
+}
+
+var _ recordings.RecordingScopeService = (*combinedService)(nil)
+
+func recordingScopeIssuer(service *combinedService) string {
+	digest := sha256.Sum256([]byte(fmt.Sprintf("recordings-root:%p", service)))
+	return hex.EncodeToString(digest[:16])
+}
+
+func (service *combinedService) newRecordingScope() recordings.RecordingScopeRef {
+	service.scopeMu.Lock()
+	defer service.scopeMu.Unlock()
+	service.nextScopeID++
+	identity := fmt.Sprintf("%s:%d", service.scopeIssuer, service.nextScopeID)
+	digest := sha256.Sum256([]byte("recording-scope:" + identity))
+	ref, _ := (recordings.RecordingScopeRef{}).Parse(
+		service.scopeIssuer + "." + hex.EncodeToString(digest[:16]),
+	)
+	return ref
+}
+
+func (service *combinedService) recordingScope(
+	ref recordings.RecordingScopeRef,
+) (*recordingScopeBinding, error) {
+	if ref.IsZero() {
+		return nil, recordings.ErrRecordingScopeInvalid
+	}
+	issuer, token, ok := strings.Cut(ref.String(), ".")
+	if !ok || !validRecordingScopeToken(issuer) || !validRecordingScopeToken(token) {
+		return nil, recordings.ErrRecordingScopeInvalid
+	}
+	if issuer != service.scopeIssuer {
+		return nil, recordings.ErrRecordingScopeForeign
+	}
+	service.scopeMu.RLock()
+	binding, ok := service.scopeByRef[ref]
+	service.scopeMu.RUnlock()
+	if !ok {
+		return nil, recordings.ErrRecordingScopeStale
+	}
+	return binding, nil
+}
+
+func validRecordingScopeToken(value string) bool {
+	if len(value) != 32 || value == strings.Repeat("0", 32) {
+		return false
+	}
+	_, err := hex.DecodeString(value)
+	return err == nil
+}
+
+func (service *combinedService) BeginRecordingScope(
+	ctx context.Context,
+	request recordings.BeginRecordingScopeRequest,
+) (result recordings.BeginRecordingScopeResult, err error) {
+	operation := service.startOperationLog("recording_scope.begin", recordings.RecordingScopeRef{}, request.Scope)
+	defer func() { operation.finish(err) }()
+	if err := recordingScopeContext(ctx).Err(); err != nil {
+		return recordings.BeginRecordingScopeResult{}, err
+	}
+	if !request.Enabled {
+		return recordings.BeginRecordingScopeResult{}, nil
+	}
+	ref := service.newRecordingScope()
+	started, err := service.StartRecording(recordings.StartRecordingRequest{
+		Enabled:       true,
+		RecordingID:   scopeRecordingID(ref, request.RecordingID),
+		Scope:         request.Scope,
+		Target:        request.Target,
+		FlushInterval: request.FlushInterval,
+	})
+	if err != nil {
+		return recordings.BeginRecordingScopeResult{}, err
+	}
+	binding := &recordingScopeBinding{
+		recordingID: started.Status.RecordingID,
+		eventScope:  started.Status.Scope,
+		replayPlans: make(map[recordings.ReplayPlanHandle]struct{}),
+	}
+	service.scopeMu.Lock()
+	service.scopeByRef[ref] = binding
+	service.scopeMu.Unlock()
+	if err := recordingScopeContext(ctx).Err(); err != nil {
+		return recordings.BeginRecordingScopeResult{}, service.abandonRecordingScope(ref, binding, err)
+	}
+	status, err := service.scopeStatus(ref, binding)
+	if err != nil {
+		return recordings.BeginRecordingScopeResult{}, service.abandonRecordingScope(ref, binding, err)
+	}
+	return recordings.BeginRecordingScopeResult{Scope: ref, Status: status}, nil
+}
+
+func (service *combinedService) abandonRecordingScope(
+	ref recordings.RecordingScopeRef,
+	binding *recordingScopeBinding,
+	cause error,
+) error {
+	service.scopeMu.Lock()
+	if current, ok := service.scopeByRef[ref]; ok && current == binding {
+		delete(service.scopeByRef, ref)
+	}
+	service.scopeMu.Unlock()
+	_, finishErr := service.FinishRecording(recordings.FinishRecordingRequest{
+		RecordingID: binding.recordingID,
+		FinishedAt:  service.recordingFinishedAt(),
+	})
+	return errors.Join(cause, finishErr)
+}
+
+func scopeRecordingID(
+	ref recordings.RecordingScopeRef,
+	requested recordings.RecordingID,
+) recordings.RecordingID {
+	digest := sha256.Sum256([]byte("recording-id:" + string(requested) + ":" + ref.String()))
+	return recordings.RecordingID("scope-" + hex.EncodeToString(digest[:16]))
+}
+
+func (service *combinedService) AppendRecordingScopeEvent(
+	ctx context.Context,
+	request recordings.AppendRecordingScopeEventRequest,
+) (result recordings.AppendRecordingScopeEventResult, err error) {
+	// Event append is a high-volume path; lifecycle and query operations emit
+	// per-operation logs while append remains observable through its canonical
+	// event and terminal status projections.
+	if err := recordingScopeContext(ctx).Err(); err != nil {
+		return recordings.AppendRecordingScopeEventResult{}, err
+	}
+	binding, err := service.recordingScope(request.Scope)
+	if err != nil {
+		return recordings.AppendRecordingScopeEventResult{}, err
+	}
+	binding.mu.Lock()
+	defer binding.mu.Unlock()
+	service.recordingMu.Lock()
+	defer service.recordingMu.Unlock()
+	if err := recordingScopeContext(ctx).Err(); err != nil {
+		return recordings.AppendRecordingScopeEventResult{}, err
+	}
+	if binding.closed {
+		return recordings.AppendRecordingScopeEventResult{}, recordings.ErrRecordingScopeClosed
+	}
+	if binding.finalized {
+		return recordings.AppendRecordingScopeEventResult{}, errors.Join(
+			recordings.ErrRecordingScopeFinalized,
+			recordings.ErrRecordingWriteRejected,
+		)
+	}
+	current, err := service.Service.QueryRecordingStatus(recordings.RecordingStatusRequest{
+		RecordingID: binding.recordingID,
+	})
+	if err != nil {
+		return recordings.AppendRecordingScopeEventResult{}, err
+	}
+	if !validRecordingScopeEvent(current.Status, binding.eventScope, request.Event) {
+		return recordings.AppendRecordingScopeEventResult{}, recordings.ErrInvalidRecordingEvent
+	}
+	var lifecycleResult recordings.RecordRecordingEventResult
+	accepted, err := service.canonicalLedger.AppendWithValidation(
+		recordings.AppendRecordedEventRequest{Event: request.Event},
+		func(event recordings.CanonicalEvent) error {
+			if event.ID == "" {
+				return recordings.ErrInvalidAppendEvent
+			}
+			if !validRecordingScopeEvent(current.Status, binding.eventScope, event) {
+				return recordings.ErrInvalidRecordingEvent
+			}
+			var recordErr error
+			lifecycleResult, recordErr = service.Service.RecordRecordingEvent(recordings.RecordRecordingEventRequest{
+				RecordingID: binding.recordingID,
+				Event:       event,
+			})
+			return recordErr
+		},
+	)
+	if err != nil {
+		return recordings.AppendRecordingScopeEventResult{}, err
+	}
+	event := accepted.Event
+	status := scopeStatusFrom(request.Scope, lifecycleResult.Status)
+	return recordings.AppendRecordingScopeEventResult{Event: event, Status: status}, nil
+}
+
+func (service *combinedService) StartRecording(
+	request recordings.StartRecordingRequest,
+) (recordings.StartRecordingResult, error) {
+	service.recordingMu.Lock()
+	defer service.recordingMu.Unlock()
+	return service.Service.StartRecording(request)
+}
+
+func (service *combinedService) RecordRecordingEvent(
+	request recordings.RecordRecordingEventRequest,
+) (recordings.RecordRecordingEventResult, error) {
+	service.recordingMu.Lock()
+	defer service.recordingMu.Unlock()
+	return service.Service.RecordRecordingEvent(request)
+}
+
+func (service *combinedService) FinishRecording(
+	request recordings.FinishRecordingRequest,
+) (recordings.FinishRecordingResult, error) {
+	service.recordingMu.Lock()
+	defer service.recordingMu.Unlock()
+	return service.Service.FinishRecording(request)
+}
+
+func validRecordingScopeEvent(
+	status recordings.RecordingStatusFacts,
+	expectedScope recordings.CanonicalEventScope,
+	event recordings.CanonicalEvent,
+) bool {
+	if !canonical.ValidAppendEvent(event) || event.Scope != expectedScope ||
+		event.Sequence < 0 ||
+		event.Cursor.StreamGenerationID == "" || event.Cursor.Sequence != event.Sequence {
+		return false
+	}
+	if status.LastEvent == nil {
+		return true
+	}
+	return event.Cursor.StreamGenerationID == status.LastEvent.StreamGenerationID &&
+		event.Sequence > status.LastEvent.Sequence
+}
+
+func (service *combinedService) FlushRecordingScope(
+	ctx context.Context,
+	request recordings.FlushRecordingScopeRequest,
+) (result recordings.FlushRecordingScopeResult, err error) {
+	operation := service.startOperationLog("recording_scope.flush", request.Scope, recordings.CanonicalEventScope{})
+	defer func() { operation.finish(err) }()
+	if err := recordingScopeContext(ctx).Err(); err != nil {
+		return recordings.FlushRecordingScopeResult{}, err
+	}
+	binding, err := service.recordingScope(request.Scope)
+	if err != nil {
+		return recordings.FlushRecordingScopeResult{}, err
+	}
+	binding.mu.Lock()
+	defer binding.mu.Unlock()
+	if binding.closed {
+		return recordings.FlushRecordingScopeResult{}, recordings.ErrRecordingScopeClosed
+	}
+	flushed, err := service.FlushRecording(recordings.FlushRecordingRequest{
+		RecordingID: binding.recordingID,
+	})
+	if err != nil {
+		return recordings.FlushRecordingScopeResult{}, err
+	}
+	return recordings.FlushRecordingScopeResult{
+		Status: scopeStatusFrom(request.Scope, flushed.Status),
+	}, nil
+}
+
+func (service *combinedService) FinalizeRecordingScope(
+	ctx context.Context,
+	request recordings.FinalizeRecordingScopeRequest,
+) (result recordings.FinalizeRecordingScopeResult, err error) {
+	operation := service.startOperationLog("recording_scope.finalize", request.Scope, recordings.CanonicalEventScope{})
+	defer func() { operation.finish(err) }()
+	binding, err := service.recordingScope(request.Scope)
+	if err != nil {
+		return recordings.FinalizeRecordingScopeResult{}, err
+	}
+	binding.mu.Lock()
+	defer binding.mu.Unlock()
+	if binding.closed {
+		return recordings.FinalizeRecordingScopeResult{}, recordings.ErrRecordingScopeClosed
+	}
+	status, finalizeErr := service.finalizeRecordingScopeLocked(
+		ctx,
+		request.Scope,
+		binding,
+		request.FinishedAt,
+	)
+	return recordings.FinalizeRecordingScopeResult{Status: status}, finalizeErr
+}
+
+func (service *combinedService) finalizeRecordingScopeLocked(
+	ctx context.Context,
+	ref recordings.RecordingScopeRef,
+	binding *recordingScopeBinding,
+	finishedAt time.Time,
+) (recordings.RecordingScopeStatus, error) {
+	if binding.finalized {
+		return binding.terminal, binding.finalizeErr
+	}
+	if err := recordingScopeContext(ctx).Err(); err != nil {
+		return recordings.RecordingScopeStatus{}, err
+	}
+	_, err := service.FinishRecording(recordings.FinishRecordingRequest{
+		RecordingID: binding.recordingID,
+		FinishedAt:  finishedAt,
+	})
+	binding.finalized = true
+	binding.finalizeErr = err
+	status, statusErr := service.scopeStatus(ref, binding)
+	if statusErr != nil {
+		return recordings.RecordingScopeStatus{}, errors.Join(err, statusErr)
+	}
+	binding.terminal = status
+	return status, errors.Join(err, statusErr)
+}
+
+func (service *combinedService) CloseRecordingScope(
+	ctx context.Context,
+	request recordings.CloseRecordingScopeRequest,
+) (result recordings.CloseRecordingScopeResult, err error) {
+	operation := service.startOperationLog("recording_scope.close", request.Scope, recordings.CanonicalEventScope{})
+	defer func() { operation.finish(err) }()
+	binding, err := service.recordingScope(request.Scope)
+	if err != nil {
+		return recordings.CloseRecordingScopeResult{}, err
+	}
+	binding.mu.Lock()
+	defer binding.mu.Unlock()
+	if binding.closed {
+		return recordings.CloseRecordingScopeResult{
+			Scope: request.Scope, Closed: true, Status: binding.terminal,
+		}, nil
+	}
+	status, finalizeErr := service.finalizeRecordingScopeLocked(
+		ctx,
+		request.Scope,
+		binding,
+		request.FinishedAt,
+	)
+	if !binding.finalized {
+		return recordings.CloseRecordingScopeResult{Status: status}, finalizeErr
+	}
+	binding.closed = true
+	return recordings.CloseRecordingScopeResult{
+		Scope: request.Scope, Closed: true, Status: status,
+	}, finalizeErr
+}
+
+func (service *combinedService) QueryRecordingScope(
+	ctx context.Context,
+	request recordings.QueryRecordingScopeRequest,
+) (result recordings.QueryRecordingScopeResult, err error) {
+	operation := service.startOperationLog("recording_scope.status", request.Scope, recordings.CanonicalEventScope{})
+	defer func() { operation.finish(err) }()
+	if err := recordingScopeContext(ctx).Err(); err != nil {
+		return recordings.QueryRecordingScopeResult{}, err
+	}
+	binding, err := service.recordingScope(request.Scope)
+	if err != nil {
+		return recordings.QueryRecordingScopeResult{}, err
+	}
+	binding.mu.Lock()
+	defer binding.mu.Unlock()
+	if binding.closed {
+		return recordings.QueryRecordingScopeResult{}, recordings.ErrRecordingScopeClosed
+	}
+	if binding.finalized {
+		return recordings.QueryRecordingScopeResult{Status: binding.terminal}, nil
+	}
+	status, err := service.scopeStatus(request.Scope, binding)
+	if err != nil {
+		return recordings.QueryRecordingScopeResult{}, err
+	}
+	return recordings.QueryRecordingScopeResult{Status: status}, nil
+}
+
+func recordingScopeContext(ctx context.Context) context.Context {
+	if ctx == nil {
+		return context.Background()
+	}
+	return ctx
+}
+
+func (service *combinedService) scopeStatus(
+	ref recordings.RecordingScopeRef,
+	binding *recordingScopeBinding,
+) (recordings.RecordingScopeStatus, error) {
+	result, err := service.QueryRecordingStatus(recordings.RecordingStatusRequest{
+		RecordingID: binding.recordingID,
+	})
+	if err != nil {
+		return recordings.RecordingScopeStatus{}, err
+	}
+	return scopeStatusFrom(ref, result.Status), nil
+}
+
+func scopeStatusFrom(
+	ref recordings.RecordingScopeRef,
+	status recordings.RecordingStatusFacts,
+) recordings.RecordingScopeStatus {
+	result := recordings.RecordingScopeStatus{
+		Scope:          ref,
+		EventScope:     status.Scope,
+		Artifact:       status.Artifact,
+		State:          status.State,
+		AcceptedEvents: status.AcceptedEvents,
+		Failures:       append([]recordings.RecordingFailure(nil), status.Failures...),
+	}
+	if status.LastEvent != nil {
+		cursor := *status.LastEvent
+		result.LastEvent = &cursor
+	}
+	if status.FlushedThrough != nil {
+		cursor := *status.FlushedThrough
+		result.FlushedThrough = &cursor
+	}
+	if status.FinalizedAt != nil {
+		finishedAt := *status.FinalizedAt
+		result.FinalizedAt = &finishedAt
+	}
+	return result
 }
