+++ pkg/services/recordings/internal/lifecycle_runtime_recorder_boundary_test.go
@@ -1,12 +1,14 @@
 package internal
 
 import (
+	"context"
+	"errors"
+	factorydefinitions "github.com/portpowered/infinite-you/pkg/services/factory_definitions"
+	factoryruntime "github.com/portpowered/infinite-you/pkg/services/factory_runtime"
+	"github.com/portpowered/infinite-you/pkg/services/recordings"
 	"strings"
 	"testing"
 	"time"
-
-	factoryruntime "github.com/portpowered/infinite-you/pkg/services/factory_runtime"
-	"github.com/portpowered/infinite-you/pkg/services/recordings"
 )
 
 const (
@@ -97,3 +99,205 @@ func TestLifecycleRuntimeRecorderAcceptsRuntimeRootFinishedEvent(t *testing.T) {
 		t.Fatalf("finished event payload = %q, want completed state", finishedEvent.Payload)
 	}
 }
+
+func TestRuntimeRootKeepsConcurrentLedgersIsolatedAndReleasesRoutes(t *testing.T) {
+	root := NewRuntimeRoot(nil, nil, nil, nil, nil, nil, nil, nil, nil)
+	if root == nil {
+		t.Fatal("NewRuntimeRoot() returned nil")
+	}
+	topology := runtimeOpeningTopology{}
+	now := func() time.Time { return time.Unix(1_700_000_000, 0).UTC() }
+
+	first, err := root.OpenRuntime(context.Background(), recordings.RuntimeScopeRequest{
+		Topology:         topology,
+		Now:              now,
+		RecordingID:      "recording-one",
+		FactorySessionID: "session-one",
+	})
+	if err != nil {
+		t.Fatalf("OpenRuntime(first): %v", err)
+	}
+	second, err := root.OpenRuntime(context.Background(), recordings.RuntimeScopeRequest{
+		Topology:         topology,
+		Now:              now,
+		RecordingID:      "recording-two",
+		FactorySessionID: "session-two",
+	})
+	if err != nil {
+		t.Fatalf("OpenRuntime(second): %v", err)
+	}
+	if first.Ledger == second.Ledger {
+		t.Fatal("OpenRuntime returned the same ledger for concurrent sessions")
+	}
+
+	first.Ledger.RecordRunRequest()
+	if got := len(first.Ledger.CanonicalEvents()); got != 1 {
+		t.Fatalf("first ledger events = %d, want 1", got)
+	}
+	if got := len(second.Ledger.CanonicalEvents()); got != 0 {
+		t.Fatalf("second ledger events = %d, want 0 before its own append", got)
+	}
+
+	finishedAt := now().Add(time.Second)
+	if err := first.Recorder.Finalize(finishedAt); err != nil {
+		t.Fatalf("Finalize(first): %v", err)
+	}
+	if err := first.Recorder.Finalize(finishedAt.Add(time.Second)); err != nil {
+		t.Fatalf("Finalize(first) second call: %v", err)
+	}
+	if _, err := root.SubscribeFrom(context.Background(), recordings.SubscribeRequest{
+		Scope: recordings.CanonicalEventScope{FactorySessionID: "session-one"},
+	}); !errors.Is(err, recordings.ErrReconnectCursorUnavailable) {
+		t.Fatalf("Subscribe(closed first session) = %v, want isolated route failure", err)
+	}
+
+	second.Ledger.RecordRunRequest()
+	if got := len(second.Ledger.CanonicalEvents()); got != 1 {
+		t.Fatalf("second ledger events = %d, want 1", got)
+	}
+	if err := second.Recorder.Finalize(finishedAt); err != nil {
+		t.Fatalf("Finalize(second): %v", err)
+	}
+}
+
+func TestRuntimeRootActiveRecordingOwnsOpaqueScopeAndFinalizesOnce(t *testing.T) {
+	root, opened, now := openActiveRuntime(t)
+	if opened.Scope.IsZero() {
+		t.Fatal("OpenRuntime(active) returned a zero scope")
+	}
+	assertActiveRecordingStarted(t, root)
+	opened.Ledger.RecordRunRequest()
+	scopeStatus := queryActiveScope(t, root, opened.Scope)
+	appendActiveScopeEvent(t, root, opened.Scope, scopeStatus)
+	finalizeActiveRuntime(t, opened.Recorder, now)
+	assertActiveRecordingFinalized(t, root)
+	assertActiveScopeClosed(t, root, opened.Scope)
+}
+
+func openActiveRuntime(t *testing.T) (recordings.Root, recordings.RuntimeScopeResult, func() time.Time) {
+	t.Helper()
+	snapshot, err := factorydefinitions.NewFactorySnapshot(map[string]any{
+		"id": "runtime-opening-test",
+	})
+	if err != nil {
+		t.Fatalf("NewFactorySnapshot: %v", err)
+	}
+	root := NewRuntimeRoot(
+		nil,
+		nil,
+		nil,
+		func(
+			factorydefinitions.FactorySnapshotSource,
+			string,
+			map[string]string,
+		) (*factorydefinitions.FactorySnapshot, error) {
+			return snapshot, nil
+		},
+		nil,
+		nil,
+		nil,
+		nil,
+	)
+	now := func() time.Time { return time.Unix(1_700_000_100, 0).UTC() }
+	opened, err := root.OpenRuntime(context.Background(), recordings.RuntimeScopeRequest{
+		Topology:         runtimeOpeningTopology{},
+		Now:              now,
+		RecordingID:      "recording-active",
+		RecordPath:       "recording.json",
+		FactorySessionID: "session-active",
+	})
+	if err != nil {
+		t.Fatalf("OpenRuntime(active): %v", err)
+	}
+	return root, opened, now
+}
+
+func assertActiveRecordingStarted(t *testing.T, root recordings.Root) {
+	t.Helper()
+	status, err := root.QueryRecordingStatus(recordings.RecordingStatusRequest{
+		RecordingID: "recording-active",
+	})
+	if err != nil {
+		t.Fatalf("QueryRecordingStatus(active): %v", err)
+	}
+	if status.Status.AcceptedEvents != 1 {
+		t.Fatalf("active recording events = %d, want initial snapshot event", status.Status.AcceptedEvents)
+	}
+}
+
+func queryActiveScope(
+	t *testing.T,
+	root recordings.Root,
+	scope recordings.RecordingScopeRef,
+) recordings.QueryRecordingScopeResult {
+	t.Helper()
+	scopeStatus, err := root.QueryRecordingScope(context.Background(), recordings.QueryRecordingScopeRequest{
+		Scope: scope,
+	})
+	if err != nil || scopeStatus.Status.LastEvent == nil {
+		t.Fatalf("QueryRecordingScope(active) = (%#v, %v), want initial cursor", scopeStatus, err)
+	}
+	return scopeStatus
+}
+
+func appendActiveScopeEvent(
+	t *testing.T,
+	root recordings.Root,
+	scope recordings.RecordingScopeRef,
+	scopeStatus recordings.QueryRecordingScopeResult,
+) {
+	t.Helper()
+	nextSequence := scopeStatus.Status.LastEvent.Sequence + 1
+	scopeEvent := scopedScopeEvent("runtime-scope-event", nextSequence, scopeStatus.Status.EventScope)
+	scopeEvent.Cursor.StreamGenerationID = scopeStatus.Status.LastEvent.StreamGenerationID
+	appended, err := root.AppendRecordingScopeEvent(context.Background(), recordings.AppendRecordingScopeEventRequest{
+		Scope: scope,
+		Event: scopeEvent,
+	})
+	if err != nil || appended.Status.AcceptedEvents != 2 {
+		t.Fatalf("AppendRecordingScopeEvent(active) = (%#v, %v), want second accepted event", appended, err)
+	}
+}
+
+func finalizeActiveRuntime(t *testing.T, recorder recordings.RuntimeRecorder, now func() time.Time) {
+	t.Helper()
+	finishedAt := now().Add(time.Second)
+	if err := recorder.Finalize(finishedAt); err != nil {
+		t.Fatalf("Finalize(active): %v", err)
+	}
+	if err := recorder.Finalize(finishedAt.Add(time.Second)); err != nil {
+		t.Fatalf("Finalize(active) second call: %v", err)
+	}
+}
+
+func assertActiveRecordingFinalized(t *testing.T, root recordings.Root) {
+	t.Helper()
+	status, err := root.QueryRecordingStatus(recordings.RecordingStatusRequest{
+		RecordingID: "recording-active",
+	})
+	if err != nil {
+		t.Fatalf("QueryRecordingStatus(finalized): %v", err)
+	}
+	if status.Status.State != recordings.RecordingFinalized || status.Status.AcceptedEvents != 3 {
+		t.Fatalf("finalized active recording = %#v, want FINALIZED with initial, scoped, and terminal events", status.Status)
+	}
+}
+
+func assertActiveScopeClosed(t *testing.T, root recordings.Root, scope recordings.RecordingScopeRef) {
+	t.Helper()
+	if _, err := root.QueryRecordingScope(context.Background(), recordings.QueryRecordingScopeRequest{
+		Scope: scope,
+	}); !errors.Is(err, recordings.ErrRecordingScopeClosed) {
+		t.Fatalf("QueryRecordingScope(closed): %v, want closed-scope error", err)
+	}
+}
+
+type runtimeOpeningTopology struct{}
+
+func (runtimeOpeningTopology) RecordingInitialStructure(
+	...factorydefinitions.RuntimeDefinitionLookup,
+) recordings.InitialStructurePayload {
+	return recordings.InitialStructurePayload{}
+}
+
+var _ recordings.InitialStructureSource = runtimeOpeningTopology{}
